refactor: extracted addExercise component
