+++ src/App.css
@@ -1,6 +1,1 @@
-.App-entry {
-  padding: 50px;
-  textAlign: "left";
-  maxWidth: 500px;
-  margin: "auto"
-}
+
