+++ src/addExercise.css
@@ -1,0 +1,6 @@
+.App-entry {
+  padding: 50px;
+  textAlign: "left";
+  maxWidth: 500px;
+  margin: "auto"
+}
